requirements.txt
anyio==3.4.0
certifi==2021.10.8
charset-normalizer==2.0.9
fastapi==0.70.0
idna==3.3
kafka-python==2.0.2
pydantic==1.8.2
requests==2.26.0
sniffio==1.2.0
starlette==0.16.0
typing-extensions==4.0.1
urllib3==1.26.7
python-snappy==0.7.1
